Rename Comment to Message 2
