--- src/domain/shared/components/Comments/models/comment.ts
@@ -1,8 +1,0 @@
-import { Author } from './author';
-
-export interface Comment {
-  id: string;
-  author?: Author;
-  createdAt: Date;
-  body: string;
-}
